Make sure that QtWebKit can find the libraries it needs on Windows.
Looking at the code, it looks like the CMake build is broken on Windows
since trying to build QtWebKit as mentioned at
https://github.com/annulen/webkit/wiki/Building-QtWebKit-on-Windows
will actually not succeed in downloading the libraries QtWebKit needs.
In fact, even if it did succeed, those libraries wouldn’t be found by
the CMake script. Anyway, we download a smaller version of those
libraries ourselves. Smaller because the ‘official’ libraries include
32-bit versions of the libraries, which we don’t need, as well as the
WebP library, which we skip (for now, at least).